gosling::context

Enum ContextEvent

pub enum ContextEvent {

    TorBootstrapStatusReceived {
        progress: u32,
        tag: String,
        summary: String,
    },
    TorBootstrapCompleted,
    TorLogReceived {
        line: String,
    },
    IdentityClientChallengeReceived {
        handle: HandshakeHandle,
        endpoint_challenge: Document,
    },
    IdentityClientHandshakeCompleted {
        handle: HandshakeHandle,
        identity_service_id: V3OnionServiceId,
        endpoint_service_id: V3OnionServiceId,
        endpoint_name: String,
        client_auth_private_key: X25519PrivateKey,
    },
    IdentityClientHandshakeFailed {
        handle: HandshakeHandle,
        reason: Error,
    },
    IdentityServerPublished,
    IdentityServerHandshakeStarted {
        handle: HandshakeHandle,
    },
    IdentityServerEndpointRequestReceived {
        handle: HandshakeHandle,
        client_service_id: V3OnionServiceId,
        requested_endpoint: String,
    },
    IdentityServerChallengeResponseReceived {
        handle: HandshakeHandle,
        challenge_response: Document,
    },
    IdentityServerHandshakeCompleted {
        handle: HandshakeHandle,
        endpoint_private_key: Ed25519PrivateKey,
        endpoint_name: String,
        client_service_id: V3OnionServiceId,
        client_auth_public_key: X25519PublicKey,
    },
    IdentityServerHandshakeRejected {
        handle: HandshakeHandle,
        client_allowed: bool,
        client_requested_endpoint_valid: bool,
        client_proof_signature_valid: bool,
        client_auth_signature_valid: bool,
        challenge_response_valid: bool,
    },
    IdentityServerHandshakeFailed {
        handle: HandshakeHandle,
        reason: Error,
    },
    EndpointClientHandshakeCompleted {
        handle: HandshakeHandle,
        endpoint_service_id: V3OnionServiceId,
        channel_name: String,
        stream: TcpStream,
    },
    EndpointClientHandshakeFailed {
        handle: HandshakeHandle,
        reason: Error,
    },
    EndpointServerPublished {
        endpoint_service_id: V3OnionServiceId,
        endpoint_name: String,
    },
    EndpointServerHandshakeStarted {
        handle: HandshakeHandle,
    },
    EndpointServerChannelRequestReceived {
        handle: HandshakeHandle,
        client_service_id: V3OnionServiceId,
        requested_channel: String,
    },
    EndpointServerHandshakeCompleted {
        handle: HandshakeHandle,
        endpoint_service_id: V3OnionServiceId,
        client_service_id: V3OnionServiceId,
        channel_name: String,
        stream: TcpStream,
    },
    EndpointServerHandshakeRejected {
        handle: HandshakeHandle,
        client_allowed: bool,
        client_requested_channel_valid: bool,
        client_proof_signature_valid: bool,
    },
    EndpointServerHandshakeFailed {
        handle: HandshakeHandle,
        reason: Error,
    },

}

Events to signal completion of asynchronous Context operations

Variants

TorBootstrapStatusReceived

Tor bootstrap progress

Fields

progress: u32

Bootstrap percent compeletion

tag: String

A short string indicating the completed bootstrap step

summary: String

A longer human-readable summary of the bootstrap progress

TorBootstrapCompleted

Tor bootstrap completed

TorLogReceived

Human-readable logs from the Context’s TorProvider

Fields

line: String

Human-readable debug log

IdentityClientChallengeReceived

An identity client has received a challenge request from an identy server

To continue the handshake, the client must call Context::identity_client_handle_challenge_received()

Fields

handle: HandshakeHandle

The handle of the in-progress handshake

endpoint_challenge: Document

An application specific challenge object used by the identity client to create a challenge response object

IdentityClientHandshakeCompleted

An identity client has successfully completed an identity handshake and may now access the requested endpoint server.

Fields

handle: HandshakeHandle

The handle of the completed handshake

identity_service_id: V3OnionServiceId

The onion-service service-id of the identity server the client has completed an identity handshake with

endpoint_service_id: V3OnionServiceId

The onion-service service-id of the requested endpoint server

endpoint_name: String

The ASCII-encoded name of the requested endpoint server

client_auth_private_key: X25519PrivateKey

The private x25519 client-auth key required to access the requested endpoint server

IdentityClientHandshakeFailed

An incoming identit handshake has failed

Fields

handle: HandshakeHandle

The handle of the failed handshake

reason: Error

The failure reason

IdentityServerPublished

The identity server’s onion-service has been published and may be reachable by identity clients

IdentityServerHandshakeStarted

An identity server has received an incoming connection and the handshake is ready to begin

Fields

handle: HandshakeHandle

The handle of the new handshake

IdentityServerEndpointRequestReceived

An identity server has received a request for an endpoint from an identity client.

To continue the handshake, the server must call Context::identity_server_handle_endpoint_request_received()

Fields

handle: HandshakeHandle

The handle of the in-progress handshake

client_service_id: V3OnionServiceId

The alleged onion-service service-id of the connecting client

requested_endpoint: String

The ASCII-encoded name of the requested endpoint server

IdentityServerChallengeResponseReceived

An identity server has received a challenge response from an identity client.

To continue the handshake, the server must call Context::identity_server_handle_challenge_response_received()

Fields

handle: HandshakeHandle

The handle of the in-progress handshake

challenge_response: Document

An application specific challenge response object created by the identity client in response to the identity server’s challenge object

IdentityServerHandshakeCompleted

An identity server’s handshake has completed.

Fields

handle: HandshakeHandle

The handle of the completed handshake

endpoint_private_key: Ed25519PrivateKey

The ed25519 private key of requested endpoint server

endpoint_name: String

The ASCII-encoded name of the requested endpoint server

client_service_id: V3OnionServiceId

The onion-service service-id of the authenticated client

client_auth_public_key: X25519PublicKey

The public x25519 client-auth key used to encrypt the endpoint server’s onion-service descriptor

IdentityServerHandshakeRejected

An identity server has rejected an identity client’s endpoint-request.

There are multiple potential reasons why a handshake may be rejected and this event provides a breakdown on which part(s) failed specifically.

Fields

handle: HandshakeHandle

The handle of the rejected handshake

client_allowed: bool

false if the client was rejected based on their onion-service service-id

client_requested_endpoint_valid: bool

false if the requested endpoint name was not understood by the server

client_proof_signature_valid: bool

false if the client failed its authentication proof (i.e. potential attempt at identity client impersonation)

client_auth_signature_valid: bool

false if the client fails its x25519 key-ownership proof (i.e. potential attempt at use an x25519 public key not owned by the client)

challenge_response_valid: bool

false if the client’s challenge response was not suitable

IdentityServerHandshakeFailed

An incoming identity handshake has failed.

Fields

handle: HandshakeHandle

The handle of the failed handshake

reason: Error

The failure reason

EndpointClientHandshakeCompleted

An endpoint client has successfully completed an endpoint handshake and may now communicate freely with the endpoint server.

Fields

handle: HandshakeHandle

The handle of the completed handshake

endpoint_service_id: V3OnionServiceId

The onion-service service-id of the endpoint server the client has connected to

channel_name: String

The ASCII-encoded name of the requested channel on the endpoint server

stream: TcpStream

The resulting TCP connection to the endpoint server

EndpointClientHandshakeFailed

An outgoing endpoint handshake has failed.

Fields

handle: HandshakeHandle

The handle of the failed handshake

reason: Error

The failure reason

EndpointServerPublished

The endpoint server’s onion-service has been published and may be reachable by endpoint clients.

Fields

endpoint_service_id: V3OnionServiceId

The onion-service service-id of the published endpoint server

endpoint_name: String

The name of the published endpoint server

EndpointServerHandshakeStarted

An endpoint server has received an incoming connection and the handshake is ready to begin.

Fields

handle: HandshakeHandle

The handle of the new handshake

EndpointServerChannelRequestReceived

An endpoint server has received a request for a channel from an endpoint client.

To continue the handshake, the server must call Context::endpoint_server_handle_channel_request_received()

Fields

handle: HandshakeHandle

The handle of the in-progress handshake

client_service_id: V3OnionServiceId

The alleged onion-service service-id of the connecting client

requested_channel: String

The ASCII-encoded name of the requested channel

EndpointServerHandshakeCompleted

An endpoint server’s handshake has completed

Fields

handle: HandshakeHandle

The handle of the completed handshake

endpoint_service_id: V3OnionServiceId

The onion-service service-id of the endpoint server which an endpoint client has connected to

client_service_id: V3OnionServiceId

The onion-service service-id of the connected client

channel_name: String

The ASCII-encoded name of the client’s requested channel

stream: TcpStream

The resulting TCP connection to tohe endpoint clientt

EndpointServerHandshakeRejected

An endpoint server has rejected an endpoint client’s channel request.

There are multiple potential reasons why a handshake may be rejected and this event provides a breakdown on which part(s) failed specifically.

Fields

handle: HandshakeHandle

The handle of the rejected handshake

client_allowed: bool

false if the client was rejected based on their onion-service service-id

client_requested_channel_valid: bool

false if the requested channel name was not understood by the server

client_proof_signature_valid: bool

false if the client failed its authentication proof (i.e. potential attempt at endpoint client impersonation)

EndpointServerHandshakeFailed

An incoming endpoint handshake has failed.

Fields

handle: HandshakeHandle

The handle of the failed handshake

reason: Error

The failure reason